Fisher's transformation of the product moment correlation coefficient

Density, distribution function, quantile function and random generator for the distribution of Fisher's transformation of product moment correlation, based on a random sample from a bivariate normal distribution

Usage

dFcorr(x, N, rho = 0, log = FALSE)

pFcorr(q, N, rho = 0, lower.tail = TRUE, log.p = FALSE)

qFcorr(p, N, rho = 0, lower.tail = TRUE, log.p = FALSE)

rFcorr(n, N, rho = 0, lower.tail = TRUE, log.p = FALSE)

Arguments

x, q

Numeric vectors of quantiles.

N

Numeric vector. Number of observations, (N > 3).

rho

Numeric vector. Population correlations, (-1 < rho < 1).

log, log.p

A logical scalar; if TRUE, probabilities p are given as log(p).

lower.tail

A logical scalar. If TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].

p

A numeric vector of probabilities in [0,1].

n

Numeric scalar. The number of observations to be simulated. If length(n) > 1 then length(n) is taken to be the number required.

Details

These functions rely on the correlation coefficient functions in the SuppDists package. SuppDists must be installed in order for these functions to work.

References

Fisher, R. A. (1915). Frequency distribution of the values of the correlation coefficient in samples of an indefinitely large population. Biometrika, 10(4), 507-521.

Fisher, R. A. (1921). On the "probable error" of a coefficient of correlation deduced from a small sample. Metron, 1, 3-32. https://digital.library.adelaide.edu.au/dspace/bitstream/2440/15169/1/14.pdf

See also

Examples

got_SuppDists <- requireNamespace("SuppDists", quietly = TRUE)

if (got_SuppDists) {
  dFcorr(-1:1, N = 10)
  dFcorr(0, N = 11:20)

  pFcorr(0.5, N = 10)
  pFcorr(0.5, N = 10, rho = c(0, 0.3))

  qFcorr((1:9)/10, N = 10, rho = 0.2)
  qFcorr(0.5, N = c(10, 20), rho = c(0, 0.3))

  rFcorr(6, N = 10, rho = 0.6)
}
#> [1] 1.5502923 1.0848058 1.1375845 0.8172382 0.7177892 0.5587049