smovie: some movies to illustrate concepts in statistics

These movies are animations used to illustrate key statistical ideas. They are produced using the rpanel package.

Details

When one of these functions is called R opens up a small parameter window containing clickable buttons that can be used to change parameters underlying the plot. For the effects of these buttons see the documentation of the individual functions.

See vignette("smovie-vignette", package = "smovie") for an overview of the package and the user-friendly menu panel.

There are movies on the following topics.

Probability distributions

Discrete distributions

Continuous distributions

Sampling distributions

Central Limit Theorem: sampling distribution of a sample mean

Central Limit Theorem for sample quantiles: sampling distribution of the 100p% sample quantile

Sampling distributions of the mean and median

Extremal Types Theorem: sampling distribution of a sample maximum

Correlation coefficient

Regression

Leverage and influence in simple linear regression

Hypothesis testing

Wald, Wilks and Score tests

Testing simple hypotheses

References

Bowman, A., Crawford, E., Alexander, G. and Bowman, R. W. (2007). rpanel: Simple Interactive Controls for R Functions Using the tcltk Package. Journal of Statistical Software, 17(9), 1-18. doi:10.18637/jss.v017.i09 .